pyproject.toml
[project]
name = "clovers-plugin-test"
version = "0.1.0"
description = ""
authors = [{ name = "KarisAya", email = "1048827424@qq.com" }]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "toml (>=0.10.2,<0.11.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "mplfinance (>=0.12.10b0,<0.13.0)",
    "clovers-utils[all] (>=0.2.3,<0.3.0)",
    "openai (>=1.60.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "websockets (>=15.0.1,<16.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
